backend_b/requirements.txt
asgiref
Django
gunicorn
packaging
psycopg2-binary
python-decouple
sqlparse
typing_extensions
whitenoise
Django
djangorestframework
django-cors-headers